requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aioredis==2.0.1
aiosignal==1.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
black==23.9.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
daphne==4.0.0
defusedxml==0.7.1
Deprecated==1.2.14
dj-rest-auth==5.0.0
Django==4.2.5
django-allauth==0.56.1
django-cors-headers==4.2.0
django-defender==0.9.7
django-discover-runner==1.0
django-environ==0.11.2
django-ipware==5.0.0
djangochannelsrestframework==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
frozenlist==1.4.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-secret-manager==2.16.3
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.12.6
grpcio==1.58.0
grpcio-status==1.58.0
gunicorn==21.2.0
hiredis==2.2.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
phonenumbers==8.13.20
platformdirs==3.10.0
proto-plus==1.22.3
protobuf==4.24.3
psycopg2-binary==2.9.3
pwned-passwords-django==2.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.2
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
twilio==8.8.0
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
wrapt==1.15.0
yarl==1.9.2
zope.interface==6.0